requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
groq==0.4.1
langchain-groq==0.0.1
pinecone-client==2.2.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
PyPDF2==3.0.1
langchain==0.0.340
chromadb==0.4.17
sentence-transformers==2.2.2
tiktoken==0.5.2
numpy==1.24.3
pandas==2.1.4
google-generativeai==0.3.2
transformers==4.36.2
torch==2.1.2
pypdf==3.17.4
langchain-community==0.0.38
langchain-huggingface==0.0.3
